backend/requirements.txt
certifi==2020.6.20
Django==3.1.2
django-cors-headers==3.5.0
djangorestframework==3.12.1
gunicorn==20.0.4
httpcore==0.12.0
httptools==0.1.1
Markdown==3.3
petl==1.6.8
requests==2.24.0
uvicorn==0.12.1
uvloop==0.14.0
